Fix bugs and improve tests

+++ test/basicnodes.sh
@@ -11,8 +11,13 @@ makeTestDir
 
 # =====
 # Nodes
-echo 'node1 6 address=localhost' > ${QPY_MU_DIR}/nodes
-echo 'node2 6 address=localhost' >> ${QPY_MU_DIR}/nodes
+if [[ -f nodes ]]
+then
+    cp nodes ${QPY_MU_DIR}/nodes
+else
+    echo 'node1 6 address=localhost' > ${QPY_MU_DIR}/nodes
+    echo 'node2 6 address=localhost' >> ${QPY_MU_DIR}/nodes
+fi
 echo 'even' > ${QPY_MU_DIR}/distribution_rules
 
 # =====
@@ -23,21 +28,14 @@ all_users="$QPY_U1"
 # =====
 # Go!
 testqpy_multiuser start
-sleep 1
-showMUlog
-print Waiting a cicle in qpy-multiuser...
-sleep 15
-print
-testqpy_multiuser status
-print
+wait_for 10
 
 createUser ${QPY_U1}
 testqpy ${QPY_U1} restart
-sleep 2
+wait_for 2
 
-print Users have been created. Waiting a cicle...
-sleep 5
-showUlog $QPY_U1
+print Users have been created.
+wait_for 5
 
 testqpy_multiuser status
 testqpy_multiuser variables
@@ -49,8 +47,7 @@ testqpy $QPY_U1 status
 
 testqpy $QPY_U1 sub -m 0.01 hostname
 testqpy $QPY_U1 check
-print Waiting a bit...
-sleep 3
+wait_for 3
 testqpy $QPY_U1 check
 runUser $QPY_U1 cat job_1.out
 runUser $QPY_U1 cat job_1.err
@@ -61,23 +58,21 @@ testqpy $QPY_U1 sub -m 0.01 sleep 20
 testqpy $QPY_U1 sub -m 0.01 sleep 20
 testqpy $QPY_U1 sub -m 0.01 sleep 20
 testqpy $QPY_U1 check
-print Waiting a bit...
-sleep 3
+wait_for 3
 testqpy $QPY_U1 check
 testqpy $QPY_U1 status
-print Waiting 20 secs
-sleep 20
+wait_for 20
 testqpy $QPY_U1 check
 testqpy $QPY_U1 status
 
 
 # =====
 # The end
-showMUlog
+#showMUlog
 for user in $all_users
 do
     testqpy $user config
-    showUlog  $user
+#    showUlog  $user
 done
 finish_test $all_users
