+++ test/killing.sh
@@ -11,7 +11,12 @@ makeTestDir
 
 # =====
 # Nodes
-echo 'localhost 5' > ${QPY_MU_DIR}/nodes
+if [[ -f nodes ]]
+then
+    cp nodes ${QPY_MU_DIR}/nodes
+else
+    echo 'localhost 5' > ${QPY_MU_DIR}/nodes
+fi
 echo 'even' > ${QPY_MU_DIR}/distribution_rules
 
 # =====
@@ -22,17 +27,16 @@ all_users="$QPY_U1"
 # =====
 # Go!
 testqpy_multiuser start
-sleep 1
-print Waiting a cicle in qpy-multiuser...
-sleep 15
+wait_for 10
 print
 
 createUser ${QPY_U1}
 testqpy ${QPY_U1} restart
-sleep 2
+wait_for 2
 
-print User has been created. Waiting a cicle...
-sleep 5
+print User has been created.
+wait_for 5
+
 testqpy $QPY_U1 config saveMessages true
 
 print Testing killing jobs for user $QPY_U1
@@ -40,12 +44,11 @@ print
 testqpy $QPY_U1 status
 testqpy $QPY_U1 sub -m 0.01 sleep 30
 testqpy $QPY_U1 check
-print Waiting a bit...
-sleep 3
+wait_for 3
 testqpy $QPY_U1 check
-sleep 1
+wait_for 1
 testqpy $QPY_U1 kill 1
-sleep 1
+wait_for 1
 testqpy $QPY_U1 check
 testqpy $QPY_U1 sub -m 0.01 sleep 30
 testqpy $QPY_U1 sub -m 0.01 sleep 30
@@ -59,29 +62,31 @@ testqpy $QPY_U1 sub -m 0.01 sleep 30
 testqpy $QPY_U1 sub -m 0.01 sleep 30
 testqpy $QPY_U1 sub -m 0.01 sleep 30
 print
-sleep 1
+wait_for 1
 testqpy $QPY_U1 check
 print
-sleep 5
+wait_for 1
 testqpy $QPY_U1 check
 testqpy $QPY_U1 kill 3-5 7,9
 print
-sleep 3
-for i in 1 2 3 4
+wait_for 3
+
+n=4
+for i in `seq $n`
 do
-    print Waiting 10 seconds, $i/4
-    sleep 10
+    wait_for 10 $i $n
     testqpy $QPY_U1 check
 done
 
+testqpy $QPY_U1 config
 
 # =====
 # The end
-showMUlog
+#showMUlog
 for user in $all_users
 do
     testqpy $user config
-    showUlog  $user
+#    showUlog  $user
 done
 finish_test $all_users
