+++ test/nodesattr.sh
@@ -12,11 +12,16 @@ makeTestDir
 
 # =====
 # Nodes
-echo 'node1 6 address=localhost big' > ${QPY_MU_DIR}/nodes
-echo 'node2 6 address=localhost infiband' >> ${QPY_MU_DIR}/nodes
-echo 'node3 6 address=localhost big infiband' >> ${QPY_MU_DIR}/nodes
-echo 'node4 6 address=localhost' >> ${QPY_MU_DIR}/nodes
-echo 'node5 6 address=localhost' >> ${QPY_MU_DIR}/nodes
+if [[ -f nodes ]]
+then
+    cp nodes ${QPY_MU_DIR}/nodes
+else
+    echo 'node1 6 address=localhost big' > ${QPY_MU_DIR}/nodes
+    echo 'node2 6 address=localhost infiband' >> ${QPY_MU_DIR}/nodes
+    echo 'node3 6 address=localhost big infiband' >> ${QPY_MU_DIR}/nodes
+    echo 'node4 6 address=localhost' >> ${QPY_MU_DIR}/nodes
+    echo 'node5 6 address=localhost' >> ${QPY_MU_DIR}/nodes
+fi
 echo 'even' > ${QPY_MU_DIR}/distribution_rules
 
 # =====
@@ -27,125 +32,134 @@ all_users="$QPY_U1"
 # =====
 # Go!
 testqpy_multiuser start
-sleep 1
-showMUlog
-print Waiting a cicle in qpy-multiuser...
-sleep 15
+wait_for 10
 print
-testqpy_multiuser status
 
 createUser ${QPY_U1}
 testqpy ${QPY_U1} restart
-sleep 2
+wait_for 2
+testqpy $QPY_U1 config saveMessages true
 
-print User has been created. Waiting a cicle...
-sleep 5
-showUlog $QPY_U1
+print User has been created.
+wait_for 15
+testqpy_multiuser status
 
 
 printBox Changing check pattern to see the attributes of jobs in queue
-sleep 3
+wait_for 3
 testqpy $QPY_U1 config checkFMT '%j (%s):%c (on %A; wd: %d)\n'
-sleep 3
+wait_for 3
 
 
 #=========
 printBox Running some jobs without attributes:
-sleep 3
+wait_for 3
 
 testqpy $QPY_U1 sub -m 0.01 sleep 10
 testqpy $QPY_U1 sub -m 0.01 sleep 10
 testqpy $QPY_U1 sub -m 0.01 sleep 10
 testqpy $QPY_U1 sub -m 0.01 sleep 10
 testqpy $QPY_U1 sub -m 0.01 sleep 10
-for i in 1 2 3 4
+
+n=4
+for i in `seq $n`
 do
+    wait_for 2 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 2
 done
-for i in 1 2
+n=2
+for i in `seq $n`
 do
+    wait_for 10 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 10
 done
 
 
 #=========
 printBox Running some jobs with attributes: big:
-sleep 3
+wait_for 3
 
 testqpy $QPY_U1 sub -m 0.01 -a big sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a big sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a big sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a big sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a big sleep 10
-for i in 1 2 3 4
+
+n=4
+for i in `seq $n`
 do
+    wait_for 2 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 2
 done
-for i in 1 2
+n=2
+for i in `seq $n`
 do
+    wait_for 10 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 10
 done
 
 #=========
 printBox Running some jobs with attributes: "'(big)and(infiband)'":
-sleep 3
+wait_for 3
 
 testqpy $QPY_U1 sub -m 0.01 -a '(big)and(infiband)' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a '(big)and(infiband)' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a '(big)and(infiband)' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a '(big)and(infiband)' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a '(big)and(infiband)' sleep 10
-for i in 1 2 3 4
+
+n=4
+for i in `seq $n`
 do
+    wait_for 2 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 2
 done
-for i in 1 2
+n=2
+for i in `seq $n`
 do
+    wait_for 10 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 10
 done
 
 
 #=========
 printBox Running some jobs with attributes: "'not((big)or(infiband))'":
-sleep 3
+wait_for 3
 
 testqpy $QPY_U1 sub -m 0.01 -a 'not((big)or(infiband))' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'not((big)or(infiband))' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'not((big)or(infiband))' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'not((big)or(infiband))' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'not((big)or(infiband))' sleep 10
-for i in 1 2 3 4
+
+n=4
+for i in `seq $n`
 do
+    wait_for 2 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 2
 done
-for i in 1 2
+n=2
+for i in `seq $n`
 do
+    wait_for 10 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 10
 done
 
 
 #=========
 printBox Adding default attribute, AND type
 testqpy $QPY_U1 config andAttr big
-sleep 3
+wait_for 3
 testqpy $QPY_U1 config
-sleep 3
+wait_for 3
 
 testqpy $QPY_U1 sub -m 0.01 sleep 10
 testqpy $QPY_U1 sub -m 0.01 sleep 10
@@ -156,11 +170,20 @@ testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
-for i in `seq 1 10`
+
+n=4
+for i in `seq $n`
 do
+    wait_for 2 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 5
+done
+n=2
+for i in `seq $n`
+do
+    wait_for 10 $i $n
+    testqpy $QPY_U1 status
+    testqpy $QPY_U1 check
 done
 
 
@@ -168,9 +191,9 @@ done
 printBox Adding default attribute, of OR type
 testqpy $QPY_U1 config andAttr 
 testqpy $QPY_U1 config orAttr big
-sleep 3
+wait_for 3
 testqpy $QPY_U1 config
-sleep 3
+wait_for 3
 
 testqpy $QPY_U1 sub -m 0.01 sleep 10
 testqpy $QPY_U1 sub -m 0.01 sleep 10
@@ -181,21 +204,30 @@ testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
 testqpy $QPY_U1 sub -m 0.01 -a 'infiband' sleep 10
-for i in `seq 1 10`
+
+n=4
+for i in `seq $n`
 do
+    wait_for 2 $i $n
     testqpy $QPY_U1 status
     testqpy $QPY_U1 check
-    sleep 5
+done
+n=2
+for i in `seq $n`
+do
+    wait_for 10 $i $n
+    testqpy $QPY_U1 status
+    testqpy $QPY_U1 check
 done
 
 
 # =====
 # The end
-showMUlog
+#showMUlog
 for user in $all_users
 do
     testqpy $user config
-    showUlog  $user
+#    showUlog $user
 done
 finish_test $all_users
