+++ test/small.sh
@@ -12,7 +12,12 @@ makeTestDir
 
 # =====
 # Nodes
-echo 'localhost 5' > ${QPY_MU_DIR}/nodes
+if [[ -f nodes ]]
+then
+    cp nodes ${QPY_MU_DIR}/nodes
+else
+    echo 'localhost 5' > ${QPY_MU_DIR}/nodes
+fi
 echo 'even' > ${QPY_MU_DIR}/distribution_rules
 echo $QPY_U1 > ${QPY_MU_DIR}/allowed_users
 
@@ -24,21 +29,17 @@ all_users="$QPY_U1"
 # =====
 # Go!
 testqpy_multiuser start
-sleep 1
-showMUlog
-print Waiting a cicle in qpy-multiuser...
-sleep 15
+wait_for 15
 print
 testqpy_multiuser status
 print
 
 createUser ${QPY_U1}
 testqpy ${QPY_U1} restart
-sleep 2
+wait_for 2
 
-print User has been created. Waiting a cicle...
-sleep 5
-showUlog $QPY_U1
+print User has been created.
+wait_for 5
 
 testqpy_multiuser status
 testqpy_multiuser variables
@@ -50,18 +51,17 @@ testqpy $QPY_U1 status
 
 testqpy $QPY_U1 sub -m 0.01 date
 testqpy $QPY_U1 check
-print Waiting a bit...
-sleep 3
+wait_for 3
 testqpy $QPY_U1 check
 runUser $QPY_U1 cat job_1.out
 runUser $QPY_U1 cat job_1.err
 
 # =====
 # The end
-showMUlog
+#showMUlog
 for user in $all_users
 do
     testqpy $user config
-    showUlog  $user
+#    showUlog  $user
 done
 finish_test $all_users
